Feat: 메인 페이지에서 user-agent로 분기하여 버튼 생성
